# contracts/api/marketplace-central.openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Central API
  version: 2026-04-02
paths:
...
  /marketplaces/fee-schedules:
    get:
      operationId: listMarketplaceFeeSchedules
      summary: List fee schedules for a marketplace
      tags: [marketplaces]
      parameters:
        - name: marketplace_code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketplaceFeeSchedule'
        "400":
          description: marketplace_code query param missing
  /admin/fee-schedules/seed:
    post:
      operationId: seedMarketplaceFeeSchedules
      summary: Force-seed fee schedules for a marketplace
      tags: [admin]
      parameters:
        - name: marketplace_code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  marketplace_code: { type: string }
                  seeded:           { type: integer }
        "400":
          description: Unknown marketplace_code or missing param
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/fee-schedules/sync:
    post:
      operationId: syncMarketplaceFeeSchedules
      summary: Force-sync fee schedules for a marketplace
      tags: [admin]
      parameters:
        - name: marketplace_code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  marketplace_code: { type: string }
                  synced:           { type: integer }
        "400":
          description: Unknown marketplace_code or missing param
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /marketplaces/accounts:
    get:
      summary: List marketplace accounts
      operationId: listMarketplaceAccounts
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketplaceAccountsResponse'
    post:
      summary: Create marketplace account
      operationId: createMarketplaceAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketplaceAccountRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketplaceAccount'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /marketplaces/policies:
    get:
      summary: List marketplace policies
      operationId: listMarketplacePolicies
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketplacePoliciesResponse'
    post:
      summary: Create marketplace policy
      operationId: createMarketplacePolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketplacePolicyRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketplacePolicy'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MarketplaceAccount:
      type: object
      required:
        - account_id
        - tenant_id
        - channel_code
        - marketplace_code
        - display_name
        - status
        - connection_mode
      properties:
        account_id:
          type: string
        tenant_id:
          type: string
        channel_code:
          type: string
        marketplace_code:
          type: string
        display_name:
          type: string
        status:
          type: string
        connection_mode:
          type: string
    CreateMarketplaceAccountRequest:
      type: object
      required:
        - account_id
        - channel_code
        - display_name
        - connection_mode
      properties:
        account_id:
          type: string
        channel_code:
          type: string
        display_name:
          type: string
        connection_mode:
          type: string
        marketplace_code:
          type: string
          description: "Registered marketplace plugin code (e.g. mercado_livre, shopee, magalu)."
        credentials_json:
          type: object
          additionalProperties:
            type: string
          description: "Credential key-value pairs matching credential_schema. Write-only — not returned in list responses."
    MarketplacePolicy:
      type: object
      required:
        - policy_id
        - tenant_id
        - account_id
        - commission_percent
        - fixed_fee_amount
        - default_shipping
        - tax_percent
        - min_margin_percent
        - sla_question_minutes
        - sla_dispatch_hours
        - shipping_provider
      properties:
        policy_id:
          type: string
        tenant_id:
          type: string
        account_id:
          type: string
        marketplace_code:
          type: string
          description: "The marketplace this policy applies to."
        commission_percent:
          type: number
          format: double
        commission_override:
          type: number
          format: double
          nullable: true
          description: "Tenant-specific override. Null = standard fee schedule applies."
        fixed_fee_amount:
          type: number
          format: double
        default_shipping:
          type: number
          format: double
        tax_percent:
          type: number
          format: double
          description: "Server-derived. Currently always 0."
        min_margin_percent:
          type: number
          format: double
        sla_question_minutes:
          type: integer
        sla_dispatch_hours:
          type: integer
        shipping_provider:
          type: string
          enum: [fixed, melhor_envio, marketplace]
          default: fixed
          description: "Normalized by the backend. Blank input defaults to fixed."
    CreateMarketplacePolicyRequest:
      type: object
      required:
        - policy_id
        - account_id
        - commission_percent
        - fixed_fee_amount
        - default_shipping
        - min_margin_percent
        - sla_question_minutes
        - sla_dispatch_hours
      properties:
        policy_id:
          type: string
        account_id:
          type: string
        commission_percent:
          type: number
          format: double
        commission_override:
          type: number
          format: double
          nullable: true
          description: "When set, overrides fee_schedules lookup. Null = use standard schedule."
        fixed_fee_amount:
          type: number
          format: double
        default_shipping:
          type: number
          format: double
        min_margin_percent:
          type: number
          format: double
        sla_question_minutes:
          type: integer
        sla_dispatch_hours:
          type: integer
        shipping_provider:
          type: string
          enum: [fixed, melhor_envio, marketplace]
          default: fixed
          description: "Optional. The backend trims whitespace, lowercases the value, and defaults blank/omitted input to fixed."
...
    ListMarketplaceAccountsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MarketplaceAccount'
    ListMarketplacePoliciesResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MarketplacePolicy'
...
    MarketplaceFeeSchedule:
      type: object
      properties:
        id:                 { type: string, format: uuid }
        marketplace_code:   { type: string }
        category_id:        { type: string }
        listing_type:       { type: string }
        commission_percent: { type: number }
        fixed_fee_amount:   { type: number }
        notes:              { type: string }
        source:             { type: string, enum: [api_sync, seeded, manual] }
        synced_at:          { type: string, format: date-time }
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Module-specific error code following the pattern MODULE_ENTITY_REASON.
                Known codes include: CATALOG_PRODUCT_NOT_FOUND, CATALOG_ENRICHMENT_INVALID,
                CLASSIFICATIONS_ENTITY_NOT_FOUND, CLASSIFICATIONS_DUPLICATE_NAME,
                MARKETPLACES_ACCOUNT_INVALID, MARKETPLACES_POLICY_INVALID,
                MARKETPLACES_METHOD_NOT_ALLOWED, MARKETPLACES_FEESCHEDULE_PARAM_MISSING,
                MARKETPLACES_DEFINITIONS_FETCH_FAILED, MARKETPLACES_INTERNAL_ERROR,
                PRICING_SIMULATION_INVALID, PRICING_SIMULATION_INVALID_IDENTITY,
                PRICING_SIMULATION_INVALID_BASE_PRICE, PRICING_SIMULATION_INVALID_VALUES,
                PRICING_SIMULATION_NOT_FOUND, PRICING_METHOD_NOT_ALLOWED,
                CONNECTORS_BATCH_NOT_FOUND, CONNECTORS_RETRY_MISSING_PRODUCT,
                INTERNAL_ERROR, INVALID_REQUEST, NOT_FOUND, CONFLICT.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...